\name{math}
\alias{math}
\title{Enclose a string in math dollars}
\description{Enclose a string in math dollars.}
\usage{math(x)}
\arguments{\item{x}{Character string.}}
\value{Returns x as a string within math dollars.}
\author{Kaspar Rufibach (maintainer), \email{kaspar.rufibach@gmail.com}, \cr \url{http://www.kasparrufibach.ch}}
\keyword{manip}
\keyword{character}

