\name{reporttools-package}
\alias{reporttools-package}
\alias{reporttools}
\docType{package}
\title{Functions to display descriptive statistics, and for data manipulation}
\description{
This package provides functions to generate tables of descriptive statistics for continuous and nominal variables,
as well as some further data manipulation functions. The functions in this package are especially helpful when 
writing reports of data analysis using Sweave.
}
\details{
\tabular{ll}{
Package: \tab reporttools\cr
Type: \tab Package\cr
Version: \tab 1.0.7\cr
Date: \tab 2011-11-30\cr
Depends: \tab xtable\cr
License: \tab GPL (>=2) \cr
}
At the beginning of data analysis, it is often useful to have tables of descriptive values for continuous and
nominal variables available. This package provides such functions, where the output is a LaTeX table. The 
functions are most efficiently used when generating reports combining LaTeX with R via Sweave.
}
\author{Kaspar Rufibach (maintainer), \email{kaspar.rufibach@gmail.com}, \cr \url{http://www.kasparrufibach.ch}

I thank Daniel Sabanes-Bove, Sarah Haile, and Philipp Muri for helpful discussions and remarks.
}
\references{
Rufibach, K. (2009) 
reporttools: R-Functions to Generate LaTeX Tables of Descriptive Statistics.
Journal of Statistical Software, Code Snippets, 31(1). \cr
\url{http://www.jstatsoft.org/v31/c01}.
}
\keyword{manip}
\keyword{character}
\keyword{chron}

