\name{getFonts}
\alias{getFonts}
\title{Used by the tabulating functions to format column titles}
\description{Used by the tabulating functions to format column titles.}
\usage{getFonts(font)}
\arguments{\item{font}{Provide font type.}}
\value{Returns function to format column titles.}
\author{Kaspar Rufibach (maintainer), \email{kaspar.rufibach@gmail.com}, \cr \url{http://www.kasparrufibach.ch}}
\keyword{manip}
\keyword{character}

