\name{math}
\alias{math}
\title{Enclose a string in math dollars}
\description{Enclose a string in math dollars.}
\usage{math(x)}
\arguments{\item{x}{Character string.}}
\value{Returns x as a string within math dollars.}
\author{Kaspar Rufibach (maintainer), \email{kaspar.rufibach@ifspm.uzh.ch}, \cr \url{http://www.biostat.uzh.ch/aboutus/people/rufibach.html}}
\keyword{manip}
\keyword{character}

