\name{reporttools-package}
\alias{reporttools-package}
\alias{reporttools}
\docType{package}
\title{Functions to display descriptive statistics, and for data manipulation}
\description{
This package provides functions to generate tables of descriptive statistics for continuous and nominal variables,
as well as some further data manipulation functions. The functions in this package are especially helpful when 
writing reports of data analysis using Sweave.
}
\details{
\tabular{ll}{
Package: \tab reporttools\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2008-11-13\cr
Depends: \tab xtable\cr
License: \tab GPL version 2 or newer\cr
}
At the beginning of data analysis, it is often useful to have tables of descriptive values for continuous and
nominal variables available. This package provides such functions, where the output is a Latex table. The 
functions are most efficiently used when generating reports combining Latex with R via Sweave.
}
\author{Kaspar Rufibach (maintainer), \email{kaspar.rufibach@gmail.com}}
\references{
Rufibach, K. (2008) 
reporttools: R-Functions to Generate LaTeX Tables of Descriptive Statistics.
Preprint, University of Zurich.
}
\keyword{manip}
\keyword{character}
\keyword{chron}

